general code clean up
