+++ public/index.html
@@ -53,7 +53,7 @@
       media="(prefers-color-scheme: dark)"
     />
 
-    <title>tomasmaillo</title>
+    <title>TomasMaillo.com</title>
   </head>
   <body>
     <noscript>You'll need to enable JavaScript to run this app :(</noscript>
