+++ src/Navbar/Navbar.tsx
@@ -1,13 +1,10 @@
 import { AnimatePresence, motion } from "framer-motion";
-import { FC, SetStateAction, useEffect, useState } from "react";
+import { FC, useEffect, useState } from "react";
 import { useScreenSize } from "../helpers";
-import { links, Link } from "./navbarData";
-import {
-  StyledLinkPreview,
-  StyledNavbar,
-  StyledNavbarItem,
-  StyledNavbarLink,
-} from "./styled";
+import { links } from "./navbarData";
+import NavbarItem from "./NavbarItem";
+import NavbarLogo from "./NavbarLogo";
+import { StyledNavbar } from "./styled";
 
 const container = {
   hidden: { opacity: 0 },
@@ -23,93 +20,33 @@ const item = {
   hidden: { opacity: 0 },
   show: { opacity: 1 },
 };
-interface NavbarItemProps {
-  link: Link;
-}
-const NavbarItem: FC<NavbarItemProps> = ({ link }) => {
-  const [isSmallScreen] = useScreenSize();
-  const [showPreview, setShowPreview] = useState(false);
-  const [mediaLoaded, setMediaLoaded] = useState(false);
-
-  const { text, url, preview } = link;
-
-  return (
-    <StyledNavbarItem variants={item}>
-      <StyledNavbarLink
-        href={url}
-        target="_blank"
-        onHoverStart={() => setShowPreview(true)}
-        onHoverEnd={() => setShowPreview(false)}
-      >
-        {isSmallScreen ? text.small : text.default}
-      </StyledNavbarLink>
-      <AnimatePresence>
-        {showPreview && preview && !isSmallScreen && (
-          <StyledLinkPreview
-            style={{ opacity: mediaLoaded ? 1 : 0 }}
-            initial={{ opacity: 0, top: 60 }}
-            animate={{ opacity: 1, top: 50 }}
-            exit={{ opacity: 0, transition: { duration: 0.1 } }}
-          >
-            {preview.isVideo ? (
-              <video autoPlay loop muted onLoad={() => setMediaLoaded(true)}>
-                <source src={preview.url} type="video/mp4" />
-              </video>
-            ) : (
-              <img onLoad={() => setMediaLoaded(true)} src={preview.url} />
-            )}
-          </StyledLinkPreview>
-        )}
-      </AnimatePresence>
-    </StyledNavbarItem>
-  );
-};
 
 interface NavbarProps {
   showLogo: boolean;
 }
 const Navbar: FC<NavbarProps> = ({ showLogo }) => {
-  const [scroll, setScroll] = useState(0);
-
-  const [isSmallScreen] = useScreenSize();
+  const [showNavbar, setShowNavbar] = useState(false);
 
   useEffect(() => {
-    const onScroll = (e: any) => {
-      setScroll(e.target.documentElement.scrollTop);
-    };
-    window.addEventListener("scroll", onScroll);
-
-    return () => window.removeEventListener("scroll", onScroll);
+    setTimeout(() => {
+      setShowNavbar(true);
+    }, 3700);
   }, []);
 
   return (
-    <StyledNavbar variants={container} initial="hidden" animate="show">
-      <div style={{ position: "relative", paddingRight: 40 }}>
-        <AnimatePresence>
-          {(showLogo || isSmallScreen) && (
-            <motion.div
-              style={{
-                opacity: showLogo ? 1 : 0,
-                position: "absolute",
-                top: -15,
-              }}
-              initial={{ opacity: 0, left: -90 }}
-              animate={{ opacity: 1, left: 0 }}
-              exit={{ opacity: 0, transition: { duration: 0.3 } }}
-            >
-              <img
-                src="/android-chrome-144x144.png"
-                style={{ width: 30, height: 30 }}
-              />
-            </motion.div>
-          )}
-        </AnimatePresence>
-      </div>
+    <>
+      {showNavbar && (
+        <StyledNavbar variants={container} initial="hidden" animate="show">
+          <NavbarLogo showLogo={showLogo} />
 
-      {links.map((link, i) => (
-        <NavbarItem key={i} link={link} />
-      ))}
-    </StyledNavbar>
+          {links.map((link, i) => (
+            <motion.div variants={item}>
+              <NavbarItem key={i} link={link} />
+            </motion.div>
+          ))}
+        </StyledNavbar>
+      )}
+    </>
   );
 };
 export default Navbar;
